Compass support added in place of vanilla SASS.
What is brought to us with this::
* Plenity of mixins for CSS3 and another utilities, like css sprites.
* Url management for resources like images and fonts.
* A whole framework, if necessary.
* Opt-In - You don't have to use anything the framwork provides, but you can.

Now, there will be only one SASS stylesheet non-partial, `app.scss`,
since there is no point for a Single Page Application have more than
one stylesheet if it is not loaded asynchronously. Another non-partial
will be compiled to `dist/` when performing production build, which is
not the ideal, so it must be avoided. However, when needed, the
Gruntfile should be modified to better fit this case, as put all
stylesheet under `dist/stylesheet` directory.

* Issues pointed on fixed (pull request)[https://github.com/juanghurtado/puppeteer/pull/11]